% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneuploidy_chr18}
\alias{aneuploidy_chr18}
\title{An S4 class MadSeq object}
\format{An MadSeq object}
\usage{
aneuploidy_chr18
}
\value{
MadSeq object returned from \code{\link{runMadSeq}} function, 
mitotic trisomy has been detected for the chromosome18
}
\description{
An MadSeq object returned by the function \code{\link{runMadSeq}}, 
the object contains the posterior distribution and deltaBIC value of 
a trisomy chromosome 18
}
\examples{
## to load the data
data(aneuploidy_chr18)
## check statistics of the data
summary(aneuploidy_chr18)
}
\keyword{datasets}
