% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MadSeqMethod.R
\docType{methods}
\name{deltaBIC}
\alias{deltaBIC}
\alias{deltaBIC,MadSeq-method}
\title{Accessing delta BIC of MadSeq object}
\usage{
deltaBIC(object)

\S4method{deltaBIC}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{MadSeq} object returned by \code{\link{runMadSeq}} 
function}
}
\description{
An S4 method to access the delta BIC values of \code{\link{MadSeq}} object
}
\author{
Yu Kong
}
\seealso{
\code{\link{MadSeq}}, \code{\link{runMadSeq}}
}

