% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invalid}
\alias{invalid}
\title{A dataset does not pass the requirement of MADSEQ package}
\format{A data frame with 100 rows and 4 variables:
\describe{
\item{chr}{chromosome}
\item{position}{locus of the site, in hg19}
\item{Ref_D}{read depth for reference allele}
\item{Alt_D}{read depth for alternative allele}
}}
\usage{
data(invalid)
}
\description{
An example of data that does not meet the requirement of the package.
}
\keyword{datasets}

