% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MADSEQ.R
\docType{package}
\name{MADSEQ-package}
\alias{MADSEQ-package}
\title{Mosaic Aneuploidy Detection using Sequencing Data (MADSEQ)}
\description{
The MADSEQ package provides a group of hierarchical Bayeisan models for the detection of mosaic aneuploidy,
the inference of the type of aneuploidy and also the quantification of the level of mosaic aneuploidy. These models
use Markov Chain Monte Carlo to sample the posterior distributions of parameters related to mosaic aneuploidy.
}
\details{
MADSEQ is a group of hierarchical Bayesian models used for the detection of mosaic aneuploidy. The MADSEQ model uses
the "rjags" model to sample the posterior distribution of the parameters related to mosaic aneuploidy. To fully understand
how to use the MADSEQ package, please refer to the documentation. The manual explains what data do you need, and how to process
the data to be ready for the model, what steps to follow and how to interpret the output from our model.
}
\author{
Yu Kong
}
\references{
Martyn Plummer (2015). rjags: Bayesian Graphical Models using MCMC. R package version 3-15. http://CRAN.R-project.org/package=rjags
}

