% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{MadSeqOutput}
\alias{MadSeqOutput}
\alias{as.MadSeqOutput}
\alias{is.MadSeqOutput}
\title{Produce Object of MadSeqOutput class, which is the output from runModel functions}
\usage{
MadSeqOutput(x)

as.MadSeqOutput(x)

is.MadSeqOutput(x)
}
\arguments{
\item{x}{A MadSeqOutput object}
}
\value{
An object of class MadSeqOutput.

The \code{MadSeqOutput} class has its own \code{print} method
}
\description{
The function MadSeqOutput is used to represent the output from runModel functions. A MadSeqOutput object is a list with two components:
1. posterior: the posterior estimation from the MCMC run of the model. 2. BIC: the Bayesian Information Criteria for current model.


}

