% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meiotic_trisomy}
\alias{meiotic_trisomy}
\title{A MadSeqOutput from meiotic trisomy model}
\format{An object of MadSeqOutput class, with two components:
\describe{
\item{posterior}{posteior disbribution of all parameters}
\item{BIC}{BIC value for this model}
}}
\usage{
data(meiotic_trisomy)
}
\description{
An example of MadSeqOutput from one of the MADSEQ models
}
\keyword{datasets}

