% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{modelComparison}
\alias{modelComparison}
\title{compare different MADSEQ models, and calculate deltaBIC between models.}
\usage{
modelComparison(object1, object2, ...)
}
\arguments{
\item{object1}{MadSeqOutput object, which is produced by one of the runModel functions}

\item{object2}{MadSeqOutput object}

\item{...}{MadSeqOutput object}
}
\value{
a named numeric vector indicating the deltaBIC between current model and the best model
}
\description{
compare multiple MadSeq models using BIC. Print out the deltaBIC between current model and the best model
}
\details{
the function calculated deltaBIC between the best model and given models. The model with lowest BIC value is the best model. The deltaBIC between models indicate the evidence against higher BIC
deltaBIC ~ [0,2]: Not worth more than a bare mention
deltaBIC ~ [2,6]: Positive
deltaBIC ~ [6,10]: Strong
deltaBIC >10: Very Strong
}

