% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{modelSelection}
\alias{modelSelection}
\title{select the best MADSEQ model for input data}
\usage{
modelSelection(object1, object2, ...)
}
\arguments{
\item{object1}{MadSeqOutput object, which is produced by one of the runModel functions}

\item{object2}{MadSeqOutput object}

\item{...}{MadSeqOutput object}
}
\value{
a character indicating which is the best model
}
\description{
compare multiple MadSeq models using BIC. Print out the order of preference
}
\details{
the function compare different models using BIC, the model with lowest BIC value is the best model. The deltaBIC between models indicate the evidence against higher BIC
deltaBIC ~ [0,2]: Not worth more than a bare mention
deltaBIC ~ [2,6]: Positive
deltaBIC ~ [6,10]: Strong
deltaBIC >10: Very Strong
}

