% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normalChrom}
\alias{normalChrom}
\alias{normal_coverage}
\title{Sequencing data for a normal chromosome}
\format{A data frame with 500 rows and 4 variables:
\describe{
\item{chr}{chromosome}
\item{position}{locus of the site, in hg19}
\item{Ref_D}{read depth for reference allele}
\item{Alt_D}{read depth for alternative allele}
A numeric vector with 1060 values indicating sequencing read depth for all 1060 sites.
An integer indicates the average sequencing depth for the whole genome.
}}
\usage{
data(normalChrom)

normal_coverage
}
\description{
Sequencing data for 500 heterozygous sites of one normal chormosome, and sequencing depth for all the sites from this chromosome
}
\keyword{datasets}

