% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{plotFraction}
\alias{plotFraction}
\title{plot the posterior distribution of Fraction estimate}
\usage{
plotFraction(object, prob = 0.95)
}
\arguments{
\item{object}{A MadSeqOutput object, which is produced by one of the runModel functions}

\item{prob}{A numeric number in the interval(0,1) giving the probability content of the intervals}
}
\description{
display the histogram of posterior distribution of the fraction of aneuploidy cells, with mean and HPD (highest posterior density) intervals indicated on the plot
}
\seealso{
\code{\link{plotPosterior}} \code{\link{plotMixture}}
}

