% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PlotShow.R
\docType{methods}
\name{plotMadSeq}
\alias{plotMadSeq}
\alias{plotMadSeq,MadSeq-method}
\title{density plot for posterior distribution of selected model}
\usage{
plotMadSeq(object)

\S4method{plotMadSeq}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{\link{MadSeq}} object returned by 
\code{\link{runMadSeq}} function.}
}
\value{
the density plot for parameters in the posterior distribution of 
selected model.
}
\description{
plot the density plot for each of the parameters in the posterior
distribution from selected model
}
\examples{
## load the example MadSeq object come with the package
data("aneuploidy_chr18")

## plot the posterior distribution
plotMadSeq(aneuploidy_chr18)
}
\seealso{
\code{\link{runMadSeq}}, \code{\link{plotFraction}}, 
\code{\link{plotMixture}}
}
\author{
Yu Kong

Yu Kong
}
