% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{plotMixture}
\alias{plotMixture}
\title{plot the posterior estimates of mixture disbribution}
\usage{
plotMixture(object, labels = TRUE)
}
\arguments{
\item{object}{A MadSeqOutput object, which is produced by one of the runModel functions}

\item{labels}{logical; if TRUE, will label each mixture and give the mean value on the plot}
}
\description{
display the posterior mixture disbribution of alternative allele frequencies, indicating the mean of each mixture
}
\seealso{
\code{\link{plotPosterior}} \code{\link{plotFraction}}
}

