% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Plot the posterior density estimates from runModel output}
\usage{
plotPosterior(object, variable = "all")
}
\arguments{
\item{object}{A MadSeqOutput object, which is produced by one of the runModel functions}

\item{variable}{A character vector indicates the posterior density of which variable to be plot. (default is plot all the variables)}
}
\description{
Display density plots of the posterior estimates of choosen variables, the density is produced by density function.
}

