% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MadSeqMethod.R
\docType{methods}
\name{posterior}
\alias{posterior}
\alias{posterior,MadSeq-method}
\title{Accessing posterior distribution of MadSeq object}
\usage{
posterior(object)

\S4method{posterior}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{MadSeq} object returned by \code{\link{runMadSeq}} 
function}
}
\description{
An S4 method to access the posterior distribution of
\code{\link{MadSeq}} object
}
\author{
Yu Kong

Yu Kong
}
\seealso{
\code{\link{MadSeq}}, \code{\link{runMadSeq}}
}

