% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessingShow.R
\name{prepareCoverageGC}
\alias{prepareCoverageGC}
\title{get sequencing coverage and GC content for targeted regions}
\usage{
prepareCoverageGC(target_bed, bam, genome_assembly = "hg19")
}
\arguments{
\item{target_bed}{A \code{character}, specify the path to the location of
bed file containing targeted regions.}

\item{bam}{character, path to the bam file. 
Please make sure that bam file is sorted, and the index bam is present}

\item{genome_assembly}{A \code{character}, indicating the assembly number 
of your genome. Default:"hg19". 
To see available genome_assembly, 
use \code{\link{available.genomes}} from \link{BSgenome} package}
}
\value{
a GRanges object with at least two mcols: depth and GC, 
each range indicating a targeted region
}
\description{
Given a bam file and a bed file containing targeted regions, 
return sequencing coverage and GC content for each targeted region
}
\note{
The bam file should be sorted and indexed.
}
\examples{
## specify the path to the location of bed file
target = system.file("extdata","target.bed",package="MADSEQ")

## specify the path to the bam file
aneuploidy_bam = system.file("extdata","aneuploidy.bam",package="MADSEQ")
normal_bam = system.file("extdata","normal.bam",package="MADSEQ")

## prepare coverage data for the samples
aneuploidy_cov_gc = prepareCoverageGC(target,aneuploidy_bam,"hg19")
normal_cov_gc = prepareCoverageGC(target,normal_bam,"hg19")
}
\seealso{
\code{\link{normalizeCoverage}}
}
\author{
Yu Kong
}
