% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessingShow.R
\name{prepareHetero}
\alias{prepareHetero}
\title{prepare heterozygous sites for aneuploidy detection}
\usage{
prepareHetero(vcffile, target_bed, genome = "hg19", writeToFile = TRUE,
  destination = NULL, plot = FALSE)
}
\arguments{
\item{vcffile}{A \code{character}, specify the path to the location of the 
vcf.gz file of your sample. \bold{Note:} the vcf file need to be compressed
by \code{bgzip}. The tool is part of \code{tabix} package, 
can be download from \url{http://www.htslib.org/}}

\item{target_bed}{A \code{character}, specify the path to the location of
bed file containing targeted regions.}

\item{genome}{A \code{character}, specify the assembly of your genome. 
Default: hg19. To see available genome assembly, 
use \code{\link{available.genomes}} from \link{BSgenome} package}

\item{writeToFile}{\code{Boolean} Default: \code{TRUE}. If \code{TRUE}, 
processed table containing heterozygous sites will be written to 
\code{destination} specified, the file will be named as 
"sample_filtered_heterozygous.txt". If set to \code{FALSE}, 
a \code{\link{GRanges}} object containing processed heterozygous sites
will be returned}

\item{destination}{A \code{character}, specify the path to the location 
where the processed heterozygous sites table will be written. 
Default: \code{NULL}, the file will be written to current working directory}

\item{plot}{A \code{Boolean} Default: \code{FALSE}. If \code{TRUE},
A plot showing AAF before and after filtering for problematic regions will 
be generated}
}
\value{
If \code{writeToFile} is set to TRUE, processed table will be 
written to the \code{destination}. Otherwise, a \code{\link{GRanges}}
object containing each of input sample will be returned.
}
\description{
given the vcf file and bed file containing targeted region, generate 
processed heterozygous sites for furthur analysis
}
\note{
1. The vcf file you provided need to be compressed by bgzip\cr
2. The vcf file should contain depth and allelic depth for variants in the 
FORMAT field
}
\examples{
## specify the path to the vcf.gz file for the aneuploidy sample
aneuploidy_vcf=system.file("extdata","aneuploidy.vcf.gz",package="MADSEQ")
target = system.file("extdata","target.bed",package="MADSEQ")
##------ if not write to file ------
aneuploidy_hetero=prepareHetero(aneuploidy_vcf,target,writeToFile=FALSE)

##------ if write to file ------
prepareHetero(aneuploidy_vcf, target,writeToFile=TRUE, destination=".")
}
\seealso{
\code{\link{runMadSeq}}
}
\author{
Yu Kong
}
