% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{print.MadSeqOutput}
\alias{print.MadSeqOutput}
\title{print method for MadSeqOutput object}
\usage{
\method{print}{MadSeqOutput}(x, ...)
}
\arguments{
\item{x}{An object of MadSeqOutput class}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print an object of MadSeqOutput class. It will print out the mean and 95% Highest density inverval of the posterior distribution estimated by the runModels function.
}
\seealso{
\code{\link{runModelOne}} \code{\link{runModelMonosomy}} \code{\link{runModelTrisomy}} \code{\link{runModelFour}} \code{\link{MadSeqOutput}}
}

