% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{runModelOne}
\alias{runModelOne}
\title{Use the one mixture model to fit the data}
\usage{
runModelOne(data, data_coverage, control_coverage, adapt = 10000,
  burnin = 10000, nChain = 2, nStep = 20000, thinSteps = 2,
  checkConvergence = FALSE, plot = TRUE)
}
\arguments{
\item{data}{dataframe: include at least four fields: chr (chromosome), position (position), Ref_D (read depth for reference allele) and Alt_D (read depth for alternative allele). Check if your input dataframe meets all the requirements by function validateData(data)}

\item{data_coverage}{numeric vector: the coverage information for all the sites on this chromosome (sites with all genotypes)}

\item{control_coverage}{numeric vector: control coverage information: it can be the average coverage of the whole genome. Or the mean of the normalized coverage.}

\item{adapt}{integer: the number of steps to "tune" the samplers. (default=10000)}

\item{burnin}{integer: the number of steps to "burn-in" the samplers. (default=10000)}

\item{nChain}{integer: the number of chains to run. (default=2)}

\item{nStep}{integer: total number of steps in chains to save (default=20000)}

\item{thinSteps}{interger: save data every thinSteps (default=2)}

\item{checkConvergence}{Boolean: check if the posterior estimation is converged by gelman and rubin diagnostic. Only available when you have more than 2 chains. (default=FALSE) If chains are not converged, then you should increase adapt and burnin steps.}

\item{plot}{Boolean: plot the alternative allele frequencies along the chromosome. (default=TRUE)}
}
\value{
A MadSeqOutput object with two components

\code{posterior} the posterior output from MCMC

\code{BIC} the BIC value for this model
}
\description{
Take in the heterozygous sites and coverage information, use one mixture model to fit the data
}
\seealso{
\code{\link{runModelMonosomy}} \code{\link{runModelTrisomy}} \code{\link{runModelFour}} \code{\link{validateData}}
}

