% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadSeqMethod.R
\docType{methods}
\name{summary,MadSeq-method}
\alias{summary}
\alias{summary,MadSeq-method}
\title{Summarize statistics of the MadSeq object}
\usage{
\S4method{summary}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{MadSeq} object returned by \code{\link{runMadSeq}} 
function}
}
\description{
An S4 method to summarize statistics for \code{\link{MadSeq}} object
}
\author{
Yu Kong
}

