% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{summary.MadSeqOutput}
\alias{summary.MadSeqOutput}
\title{summary statistics for MadSeqOutput object}
\usage{
\method{summary}{MadSeqOutput}(object, ...)
}
\arguments{
\item{object}{An object of MadSeqOutput class}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary.MadSeqOutput produce the summary statistics for the posterior distribution produced by runModels function.
}
\seealso{
\code{\link{runModelOne}} \code{\link{runModelMonosomy}} \code{\link{runModelTrisomy}} \code{\link{runModelFour}} \code{\link{MadSeqOutput}}
}

