% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Function.R
\name{validateData}
\alias{validateData}
\title{Check if the data is valid for MAD-SEQ analysis}
\usage{
validateData(data)
}
\arguments{
\item{data}{The data frame to be used for analysis}
}
\value{
Ture or error code indicate why the data fails the requirements
}
\description{
Take in the data frame and check if it meets all the requirements
}

