% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MadSeqMethod.R
\docType{methods}
\name{deltaBIC}
\alias{deltaBIC}
\alias{deltaBIC,MadSeq-method}
\title{Accessing delta BIC of MadSeq object}
\usage{
deltaBIC(object)

\S4method{deltaBIC}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{MadSeq} object returned by \code{\link{runMadSeq}} 
function}
}
\value{
A \code{numeric} vector containing deltaBIC values between selected
model and other models
}
\description{
An S4 method to access the delta BIC values of \code{\link{MadSeq}} object
}
\examples{
## load the example MadSeq object come with the package
data("aneuploidy_chr18")

## access deltaBIC 
deltaBIC(aneuploidy_chr18)
}
\seealso{
\code{\link{MadSeq}}, \code{\link{runMadSeq}}
}
\author{
Yu Kong
}
