% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PlotShow.R
\docType{methods}
\name{plotFraction}
\alias{plotFraction}
\alias{plotFraction,MadSeq-method}
\title{histgram for the fraction of aneuploid cells estimated by MadSeq model}
\usage{
plotFraction(object, prob = 0.95)

\S4method{plotFraction}{MadSeq}(object, prob = 0.95)
}
\arguments{
\item{object}{A \code{\link{MadSeq}} object returned by 
\code{\link{runMadSeq}} function.}

\item{prob}{A \code{numeric} value between 0~1 specify the highest posterior
interval (similar to credible interval) for the distribution. Default: 0.95.}
}
\value{
the histgram of posterior distribution of the fraction
}
\description{
histgram of the posterior distribution of the fraction of aneuploid cells 
estimated by the selected model.
}
\note{
If normal model has been selected by \code{\link{runMadSeq}} function,
no fraction plot will be produced by this function.
}
\examples{
## load the example MadSeq object come with the package
data("aneuploidy_chr18")

## plot estimated fraction of aneuploid cells
plotFraction(aneuploidy_chr18)
}
\seealso{
\code{\link{runMadSeq}}, \code{\link{plotMadSeq}}, 
\code{\link{plotMixture}}
}
\author{
Yu Kong

Yu Kong
}
