% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PlotShow.R
\docType{methods}
\name{plotMixture}
\alias{plotMixture}
\alias{plotMixture,MadSeq-method}
\title{density plot for the posterior distribution of alternative allele frequency 
estimated from the selected model}
\usage{
plotMixture(object)

\S4method{plotMixture}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{\link{MadSeq}} object returned by 
\code{\link{runMadSeq}} function.}
}
\value{
density plot for the posterior distribution of AAF
}
\description{
density plot presents the posterior distribution of alternative allele
frequency (AAF) estimated from selected model
}
\author{
Yu Kong

Yu Kong
}
\seealso{
\code{\link{runMadSeq}}, \code{\link{plotMadSeq}}, 
\code{\link{plotFraction}}
}

