% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MadSeqMethod.R
\docType{methods}
\name{posterior}
\alias{posterior}
\alias{posterior,MadSeq-method}
\title{Accessing posterior distribution of MadSeq object}
\usage{
posterior(object)

\S4method{posterior}{MadSeq}(object)
}
\arguments{
\item{object}{A \code{MadSeq} object returned by \code{\link{runMadSeq}} 
function}
}
\value{
A \code{matrix} containing posterior distribution of selected model
}
\description{
An S4 method to access the posterior distribution of
\code{\link{MadSeq}} object
}
\examples{
## load the example MadSeq object come with the package
data("aneuploidy_chr18")

## access posterior distribution
posterior(aneuploidy_chr18)
}
\seealso{
\code{\link{MadSeq}}, \code{\link{runMadSeq}}
}
\author{
Yu Kong

Yu Kong
}
